+++ CHANGELOG.md
@@ -3,7 +3,11 @@
 
 ## [Unreleased][]
 
-[Unreleased]: https://github.com/chaostoolkit/chaostoolkit-lib/compare/1.35.1...HEAD
+[Unreleased]: https://github.com/chaostoolkit/chaostoolkit-lib/compare/1.36.0...HEAD
+
+## [1.36.0][] - 2023-06-29
+
+[1.36.0]: https://github.com/chaostoolkit/chaostoolkit-lib/compare/1.35.1...1.36.0
 
 ### Added